Split roundtrip tests into their own module
