+++ LICENSE
@@ -4,8 +4,8 @@ the GPL license.
 
 
 All Python code in this library is provided under the following MIT license,
-with the exception of some files in the compatibility/ module, whose licences
-are specified within them.
+with the exception of some files in the tmxlib_test/compatibility/ module,
+whose licences are specified within them.
 
 
 Copyright (C) 2011, 2012 by Petr Viktorin
