+++ tmxlib_test/__init__.py
@@ -3,6 +3,8 @@ import os
 import pytest
 
 import tmxlib
+from tmxlib_test.compatibility.formencode_doctest_xml_compare import xml_compare
+from tmxlib.fileio import etree
 
 
 base_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data')
@@ -26,3 +28,20 @@ def file_contents(filename):
 @pytest.fixture
 def desert():
     return tmxlib.Map.open(get_test_filename('desert.tmx'))
+
+
+def assert_xml_compare(a, b):
+    report = []
+
+    def reporter(problem):
+        report.append(problem)
+
+    if not xml_compare(etree.XML(a), etree.XML(b), reporter=reporter):
+        print(a)
+        print()
+        print(b)
+        print()
+        print('XML compare report:')
+        for r_line in report:
+            print(r_line)
+        assert False
