+++ tmxlib_test/test_tmxlib.py
@@ -2,188 +2,15 @@
 from __future__ import division
 
 import os
-import sys
-import tempfile
-import json
-import collections
 
 import pytest
 
 import tmxlib
-from tmxlib.fileio import etree
-from tmxlib.compatibility.formencode_doctest_xml_compare import xml_compare
-from tmxlib_test import desert, get_test_filename, file_contents, base_path
-
-make_test_map_data = collections.namedtuple(
-    'TestMapData',
-    'filename has_gzip out_filename')(None, False, None)._replace
-
-@pytest.fixture(params=[
-        make_test_map_data(filename='desert.tmx'),
-        make_test_map_data(filename='perspective_walls.tmx'),
-        make_test_map_data(filename='sewers.tmx'),
-        make_test_map_data(filename='tilebmp-test.tmx', has_gzip=True),
-        make_test_map_data(filename='desert_nocompress.tmx'),
-        make_test_map_data(filename='desert_and_walls.tmx'),
-        make_test_map_data(filename='sewers_comment.tmx',
-                         out_filename='sewers.tmx'),
-        make_test_map_data(filename='walls_and_desert.tmx'),
-        make_test_map_data(filename='equivcheck.tmx'),
-        make_test_map_data(filename='imagelayer.tmx'),
-        make_test_map_data(filename='objects.tmx'),
-        make_test_map_data(filename='isometric_grass_and_water.tmx'
-            # NOTE: the image for this map's tileset is intentionally missing
-            ),
-    ])
-def test_map_data(request):
-    return request.param
-
-
-def assert_xml_compare(a, b):
-    report = []
-
-    def reporter(problem):
-        report.append(problem)
-
-    if not xml_compare(etree.XML(a), etree.XML(b), reporter=reporter):
-        print(a)
-        print()
-        print(b)
-        print()
-        print('XML compare report:')
-        for r_line in report:
-            print(r_line)
-        assert False
-
-
-def assert_json_safe_almost_equal(a, b, epsilon=0.00001):
-    """Test that two JSON-safe data structures are "almost equal"
-
-    Recurses into dicts, lists, tuples.
-    Checks that floats (and ints) are within epsilon from each other
-    """
-    if a == b:
-        pass
-    elif isinstance(a, (list, tuple)):
-        assert isinstance(b, (list, tuple))
-        if len(a) != len(b):
-            try:
-                raise AssertionError('Differing elements: {0}'.format(
-                    list(set(a) ^ set(b))))
-            except TypeError:
-                assert a == b
-        for i, (aa, bb) in enumerate(zip(a, b)):
-            try:
-                assert_json_safe_almost_equal(aa, bb, epsilon)
-            except:
-                print('in element {0}'.format(i))
-                raise
-    elif isinstance(a, dict):
-        assert isinstance(b, dict)
-        keys = sorted(a.keys())
-        try:
-            assert_json_safe_almost_equal(keys, sorted(b.keys()), epsilon)
-        except:
-            print('in dict keys')
-            raise
-        for key in keys:
-            try:
-                assert_json_safe_almost_equal(a[key], b[key], epsilon)
-            except:
-                print('in dict key {0!r}'.format(key))
-                raise
-    elif isinstance(a, (int, float)):
-        assert isinstance(b, (int, float))
-        assert abs(a - b) < epsilon
-    else:
-        assert a == b
+from tmxlib_test import desert, get_test_filename, file_contents
+from tmxlib_test import assert_xml_compare
 
 
 # actual test code
-def test_roundtrip_opensave(test_map_data):
-    filename, has_gzip, output_filename = test_map_data
-    if has_gzip and sys.version_info < (2, 7):
-        raise pytest.skip('Cannot test gzip on Python 2.6: missing mtime arg')
-
-    filename = get_test_filename(filename)
-    if output_filename:
-        output_filename = get_test_filename(output_filename)
-    else:
-        output_filename = filename
-    map = tmxlib.Map.open(filename)
-    for layer in map.layers:
-        # normalize mtime, for Gzip
-        layer.mtime = 0
-    temporary_file = tempfile.NamedTemporaryFile(delete=False)
-    map.check_consistency()
-    try:
-        temporary_file.close()
-        map.save(temporary_file.name)
-        assert_xml_compare(file_contents(output_filename),
-                file_contents(temporary_file.name))
-    finally:
-        os.unlink(temporary_file.name)
-
-
-def test_roundtrip_readwrite(test_map_data):
-    filename, has_gzip, output_filename = test_map_data
-    if has_gzip and sys.version_info < (2, 7):
-        raise pytest.skip('Cannot test gzip on Python 2.6: missing mtime arg')
-
-    xml = file_contents(get_test_filename(filename))
-    map = tmxlib.Map.load(xml, base_path=base_path)
-    for layer in map.layers:
-        # normalize mtime, for Gzip
-        layer.mtime = 0
-    dumped = map.dump()
-    if output_filename:
-        xml = file_contents(get_test_filename(output_filename))
-    assert_xml_compare(xml, dumped)
-
-
-def test_dict_export(test_map_data):
-    filename, has_gzip, output_filename = test_map_data
-    xml = file_contents(get_test_filename(filename))
-    map = tmxlib.Map.load(xml, base_path=base_path)
-    dct = json.load(open(get_test_filename(filename.replace('.tmx', '.json'))))
-    result = map.to_dict()
-    assert_json_safe_almost_equal(result, dct)
-
-
-def test_dict_import(test_map_data):
-    filename, has_gzip, output_filename = test_map_data
-    dct = json.load(open(get_test_filename(filename.replace('.tmx', '.json'))))
-    map = tmxlib.Map.from_dict(dct)
-
-    # Check JSON roundtrip
-
-    result_dct = map.to_dict()
-    assert_json_safe_almost_equal(result_dct, dct)
-
-    # Check XML roundtrip
-
-    if has_gzip and sys.version_info < (2, 7):
-        raise pytest.skip('Cannot test gzip on Python 2.6: missing mtime arg')
-
-    xml = file_contents(get_test_filename(filename))
-    xml_map = tmxlib.Map.load(xml, base_path=base_path)
-
-    if output_filename:
-        xml = file_contents(get_test_filename(output_filename))
-    else:
-        xml = file_contents(get_test_filename(filename))
-
-    # Have to copy presentation attrs, since those aren't in the JSON
-    for layer, xml_layer in zip(map.layers, xml_map.layers):
-        layer.compression = getattr(xml_layer, 'compression', None)
-        layer.mtime = 0
-    for tileset, xml_tileset in zip(map.tilesets, xml_map.tilesets):
-        tileset.source = xml_tileset.source
-
-    dumped = map.dump()
-    assert_xml_compare(xml, dumped)
-
-
 def test_get_layer_by_name(desert):
     assert desert.layers['Ground'].name == 'Ground'
